requirements_pyannote.txt
python-dotenv
pydub
pyannote.audio
ipykernel
speechbrain